Implemented imports, Improved name resolution, misc
The builder package now follows all imports of non standard library packages.
The builder context now holds the path of the current package. The api to invoke
builder reflects this.

The types package is now used for all name resolutions. This simplifies finding
channels, mutexes, wait groups, and functions, and enables full support for
imports. This new name resoltion in builder.processIdent() also now includes
centralised usage of the subsitutions helper.
The creation of ir.Variables has been centralized through ir.Program and
builder/variables.go. Variable indices are now unique per program and
ir.Program allows for lookup. ir.Variables also now hold a pointer to their
enclosing ir.Scope.

The builder package no longer expects users of function call results to supply
result ir.Variables. This was only used by assignments. Instead, calls now
return unnamed result variables to be assigned later.

In the builder package, all assignments (GenDecl, AssignStmt) now use one
shared method to create ir.AssignStmts. The change described above made it
possible to simplify this function.

Updated tests for imports.

+++ tests/basic/test1/test1.go
@@ -7,11 +7,11 @@ import (
 var ch chan int
 
 func testA() {
-	a, b := 42, make(chan int)
-	c, b := 65, make(chan int)
+	a, b := int32(42), make(chan int)
+	c, b := uint8(65), make(chan int)
 
-	b <- a
-	b <- c
+	b <- int(a)
+	b <- int(c)
 }
 
 func testB() {
